+++ packages/expo-ui/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unpublished
 
 - [ios] - Fix modifiers import path in the example ([#41811](https://github.com/expo/expo/pull/41811) by [@pchalupa](https://github.com/pchalupa))
+- [ios] - Add LazyHStack and LazyVStack views by [@wielski](https://github.com/wielski)
 
 ### 🛠 Breaking changes